added the averages
added the code to get original averages and then new averages from the length of the winner
BUGS: please fix the bug where D is getting a new average of 0.
IF NEEDED we can change it so we get the length of the groups from the team with the least amount of people because its easier.